+++ src/api/createCpmProduct/index.py
@@ -1,17 +1,7 @@
-from types import ModuleType
-
-from api_utils.versioning.constants import VERSIONING_STEP_ARGS
-from api_utils.versioning.steps import (
-    get_largest_possible_version,
-    get_steps_for_requested_version,
-    versioning_steps,
-)
-from domain.logging.step_decorators import logging_step_decorators
-from domain.response.steps import response_steps
+from api_utils.api_step_chain import execute_step_chain_with_location
 from event.aws.client import dynamodb_client
 from event.environment import BaseEnvironment
 from event.logging.logger import setup_logger
-from event.step_chain import StepChain
 
 from .src.v1.steps import steps as v1_steps
 
@@ -29,51 +19,8 @@ cache = {
 
 def handler(event: dict, context=None):
     setup_logger(service_name=__file__)
-    return execute_step_chain(
+    return execute_step_chain_with_location(
         event=event,
         cache=cache,
         versioned_steps=versioned_steps,
     )
-
-
-STEP_DECORATORS = [*logging_step_decorators]
-
-
-def lower_case_keys(_dict: dict[str, str]):
-    return {k.lower(): v for k, v in _dict.items()}
-
-
-def execute_step_chain(
-    event: dict, cache: dict, versioned_steps: dict[str, ModuleType]
-):
-    event["headers"] = lower_case_keys(event.get("headers", {}))
-
-    version_chain = StepChain(
-        step_chain=versioning_steps, step_decorators=STEP_DECORATORS
-    )
-    version_chain.run(
-        init={
-            VERSIONING_STEP_ARGS.EVENT: event,
-            VERSIONING_STEP_ARGS.VERSIONED_STEPS: versioned_steps,
-        }
-    )
-
-    version = None
-    location = None
-    if isinstance(version_chain.result, Exception):
-        result = version_chain.result
-    else:
-        version = version_chain.data[get_largest_possible_version]
-        steps = version_chain.data[get_steps_for_requested_version]
-        api_chain = StepChain(step_chain=steps, step_decorators=STEP_DECORATORS)
-        api_chain.run(cache=cache, init=event)
-        if isinstance(api_chain.result, Exception):
-            result = api_chain.result
-        else:
-            result, location = api_chain.result
-
-    response_chain = StepChain(
-        step_chain=response_steps, step_decorators=STEP_DECORATORS
-    )
-    response_chain.run(init=(result, version, location))
-    return response_chain.result