+++ src/layers/domain/core/device_key/tests/test_device_key_v1.py
@@ -1,11 +1,11 @@
 import pytest
 from domain.core.device_key import DeviceKeyType, validate_key
-from domain.core.error import InvalidDeviceKeyError
+from domain.core.error import InvalidKeyPattern
 
 GOOD_ID_EXAMPLES = {
     "product_id": "P.XXX-YYY",
-    "accredited_system_id": "ABC:12345",
-    "message_handling_system_id": "ABC:ABC-12345:abc123",
+    "accredited_system_id": "ABC:123456",
+    "message_handling_system_id": "ABC:ABC-123456:abc123",
 }
 
 
@@ -20,7 +20,7 @@ def test_validate_key_fail(key, type):
     if (type, key) in GOOD_ID_EXAMPLES.items():
         pytest.skip("Already covered in 'test_validate_key_pass'")
 
-    with pytest.raises(InvalidDeviceKeyError):
+    with pytest.raises(InvalidKeyPattern):
         validate_key(key=key, type=DeviceKeyType(type))
 
 
@@ -34,5 +34,5 @@ def test_validate_key_fail(key, type):
     ),
 )
 def test_validate_key_fail_other(key, type):
-    with pytest.raises(InvalidDeviceKeyError):
+    with pytest.raises(InvalidKeyPattern):
         validate_key(key=key, type=DeviceKeyType(type))