+++ src/layers/domain/repository/keys/v3.py
@@ -7,3 +7,4 @@ class TableKey(TableKeyAction, StrEnum):
     PRODUCT_TEAM = "PT"
     CPM_SYSTEM_ID = "CSI"
     CPM_PRODUCT = "P"
+    CPM_PRODUCT_KEY = "PK"